KubeOS:fix the kbimg.sh exception and pxe installation problem.
Fixed the issue of abnormal usage printing of the kbimg.sh parameter,
no verification of the -b parameter, and environment clearance in
the concurrent scenario.
Fix the problem that disks and network adapters cannot be found
due to disk and network adapter verification during PXE installation.
Fix the problem that os-proxy does not transfer imagetype and dockerimage
to the os-agent.

Signed-off-by: liyuanr <liyuanrong1@huawei.com>