bugfix: fix the problem that proxy will get all os for processing
The OS modification triggers the reconcile of the proxy. As a result, the proxy processes
the OS upgrade or configuration that does not correspond to the current node.
To solve this problem, the namedspace field is added to osinstance.spec. During each upgrade
or configuration compelets, the name and namespace of the OS corresponding to osinstance are specified.
After the upgrade or configuration is complete, the field is set to None.

Signed-off-by: liyuanr <liyuanrong1@huawei.com>