bugfix (os-operator and proxy): fix the issue that some node configurations are not delivered.
When configuring node, the operator updates osinstance and then node.
The time of the two updates is affected by the response time of the APIServer.
The update time may be different. If the proxy completes the configuration immediately
after the osinstance is updated and the node label is updated, the configuration label
on the node is not deleted. As a result, the node is skipped during the next configuration.
Therefore, the osinstance and node label check is added to the proxy.The configuration is performed
only after the operator is updated.
In addition, the logs of the operator and proxy are optimized as follows:
1. Fixe an issue where error logs are printed when the values of starttime and endtime are the same.
2. Delete the logs used during development from the time.go file.
3. The log about the successful deletion of the serial label by the operator is added.
4. Add a description before obtaining the logs of the node to be checked. (whether to add
serial labels or upgrade/configuration)
5. Logs are added when a node is being upgraded or configuration is returned in serial mode.
6. Some debug logs are added to the operator.

Signed-off-by: liyuanr <liyuanrong1@huawei.com>